Drop a lot of files and restructure tree.
The main goal here is to focus on the float64 version of the package.
To avoid all sorts of `goimports` problems, rename it as the top level
"svgof" package.

Introduce the one example I need: examples/drl2svg.go . This generates
an SVG file from a Kicad emitted drl file. I can use this SVG file to
drill PCB holes on a Snapmaker 2 A350T CNC.

Signed-off-by: Tinkerer <tinkerer@zappem.net>